# docs/07-data-api-events/specs/openapi.platform.yaml
openapi: 3.1.2
info:
  title: Korean Equity RWA PoC Platform API
  version: 1.0.0
  description: 합성 데이터만 사용하는 7단계 플랫폼 HTTP 계약이다. 상태변경 명령의 202 응답은 업무 완료가 아니다.
servers:
  - url: http://localhost:3000/api/v1
security:
  - syntheticBearer: []
...
paths:
...
  /workflows/{workflowId}:
    get:
      tags: [CommonQueries]
      operationId: getWorkflow
      summary: 업무의 독립 상태축 조회
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: 독립 상태축과 투영 완전성
          content:
            application/json:
              schema:
                $ref: './schemas/domain.schema.json#/$defs/WorkflowView'
        '404': { $ref: '#/components/responses/Error' }
...
  /primary-orders:
    get:
      tags: [CommonQueries]
      operationId: listPrimaryOrders
      summary: 1차 발행 주문 목록 조회
      responses:
        '200': { $ref: '#/components/responses/PrimaryOrderList' }
    post:
      tags: [InvestorCommands]
      operationId: createPrimaryOrder
      summary: 1차 지정가 발행 주문 접수
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas/domain.schema.json#/$defs/PrimaryOrderRequest'
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /primary-orders/{orderId}/cancellations:
    post:
      tags: [InvestorCommands]
      operationId: cancelPrimaryOrder
      summary: 1차 발행 주문 취소 요청
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reasonKo]
              properties:
                reasonKo: { type: string, minLength: 1, maxLength: 500 }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '409': { $ref: '#/components/responses/Error' }
...
  /adapter-events:
    post:
      tags: [AdapterIngress]
      operationId: receiveAdapterEvent
      summary: 서명된 모의 기관 결과 접수
      security:
        - adapterSignature: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './schemas/domain.schema.json#/$defs/AdapterEventInput' }
      responses:
        '202':
          description: 검증과 비동기 소비를 위해 접수됨. 중복이면 기존 결과를 가리킨다.
          content:
            application/json:
              schema: { $ref: './schemas/common.schema.json#/$defs/AcceptedCommand' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    syntheticBearer:
      type: http
      scheme: bearer
      bearerFormat: Synthetic-JWT
    adapterSignature:
      type: apiKey
      in: header
      name: X-Adapter-Key-Id
      description: 본문 안의 Ed25519 signature와 함께 검증한다.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' }
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
...
    WorkflowId:
      name: workflowId
      in: path
      required: true
      schema: { type: string, format: uuid }
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    PrimaryOrderList:
      description: 로컬 합성 1차 주문과 독립 상태축
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [items, projection]
            properties:
              items:
                type: array
                items: { $ref: '#/components/schemas/PrimaryOrderView' }
              projection: { $ref: './schemas/common.schema.json#/$defs/ProjectionMetadata' }
    Accepted:
      description: 비동기 업무 접수
      content:
        application/json:
          schema: { $ref: './schemas/common.schema.json#/$defs/AcceptedCommand' }
    Error:
      description: 안정적인 기계 오류와 한국어 다음 행동
      content:
        application/json:
          schema: { $ref: './schemas/common.schema.json#/$defs/ErrorResponse' }
...
  schemas:
...
    PrimaryOrderView:
      type: object
      additionalProperties: true
      required: [orderId, securityId, shareQuantity, krwLimitPrice, fundingMode, status, filledQuantity, allocatedQuantity, rightsStatus, tokenStatus, settlementStatus, simulation, projection]
      properties:
        orderId: { type: string, format: uuid }
        securityId: { const: '990001' }
        shareQuantity: { $ref: './schemas/common.schema.json#/$defs/PositiveIntegerString' }
        krwLimitPrice: { $ref: './schemas/common.schema.json#/$defs/PositiveIntegerString' }
        fundingMode: { $ref: './schemas/domain.schema.json#/$defs/PrimaryFundingMode' }
        status: { type: string, pattern: '^[A-Z][A-Z0-9_]+$' }
        filledQuantity: { $ref: './schemas/common.schema.json#/$defs/NonNegativeIntegerString' }
        allocatedQuantity: { $ref: './schemas/common.schema.json#/$defs/NonNegativeIntegerString' }
        rightsStatus: { type: string, pattern: '^[A-Z][A-Z0-9_]+$' }
        tokenStatus: { type: string, pattern: '^[A-Z][A-Z0-9_]+$' }
        settlementStatus: { type: string, pattern: '^[A-Z][A-Z0-9_]+$' }
        defaultResolution: { type: string, enum: [REPLACEMENT_SHARES, CASH_COMPENSATION] }
        cashCompensationUsdMinor: { $ref: './schemas/common.schema.json#/$defs/PositiveIntegerString' }
        quarantineReason: { type: string, minLength: 1 }
        simulation: { const: true }
        projection: { $ref: './schemas/common.schema.json#/$defs/ProjectionMetadata' }
...